refactor: clean up code for fixtures deployment
refactor: remove empty fixture files

refactor: setup roles
